cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "safeyolo"
version = "0.1.0"
description = "Security proxy CLI for AI coding agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Craig Balding", email = "craig@threatspotting.com" }
]
keywords = ["security", "proxy", "ai", "llm", "mitmproxy", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "typer>=0.9.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "jinja2>=3.0",
    "cryptography>=44.0.0",
    "pydantic>=2.0",
    "tomlkit>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
safeyolo = "safeyolo.cli:app"

[project.urls]
Homepage = "https://github.com/craigbalding/safeyolo"
Documentation = "https://github.com/craigbalding/safeyolo#readme"
Repository = "https://github.com/craigbalding/safeyolo"
Issues = "https://github.com/craigbalding/safeyolo/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/safeyolo"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.hatch.build]
include = [
    "src/safeyolo/**/*.py",
    "src/safeyolo/*.sh",
    "src/safeyolo/templates/**/*.j2",
    "src/safeyolo/templates/**/*.md",
    "src/safeyolo/templates/**/*.yaml",
    "src/safeyolo/templates/**/*.toml",
    "src/safeyolo/templates/**/.env.example",
    "src/safeyolo/templates/*.sudoers",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
# uds-networking/ holds standalone proof-of-concept runners (sys.exit at
# module top, not pytest tests). They're invoked directly by the VPS
# smoke-test scripts. Excluded from collection so pytest doesn't import
# them and trigger their top-level exits.
norecursedirs = ["uds-networking", "__pycache__"]
